Conflicts:
	src/main/java/org/marble/util/Connectors.java
	src/main/resources/org/marble/level/menu.level

+++ src/main/java/org/marble/ball/BallKind.java
@@ -5,51 +5,51 @@ package org.marble.ball;
  */
 public enum BallKind {
     /** A stone ball: heavy and slow */
-    Stone(6.0, 0.5, 0.5, 4.0),
+    Stone(6.0f, 0.5f, 0.5f, 4.0f),
 
     /** A wooden ball: light and agile */
-    Wood(1.0, 1.0, 1.0, 0.8),
+    Wood(1.0f, 1.0f, 1.0f, 0.8f),
 
     /** A fabric ball: very light and flimsy */
-    Fabric(0.25, 2.0, 2.0, 0.5),
+    Fabric(0.25f, 2.0f, 2.0f, 0.5f),
 
     /** An easily controlled ball that might break when moved too quickly */
-    Glass(2.0, 1.0, 0.5, 1.0),
+    Glass(2.0f, 1.0f, 0.5f, 1.0f),
 
     /**
      * A ball that leaves a trail of mercury as it moves, slowly growing smaller
      */
-    Mercury(8.0, 2.0, 16.0, 8.0);
+    Mercury(8.0f, 2.0f, 16.0f, 8.0f);
 
-    private static final double MASS_SCALE = 4.0;
-    private static final double STABILITY_SCALE = 4.0;
+    private static final float MASS_SCALE = 4.0f;
+    private static final float STABILITY_SCALE = 4.0f;
 
-    private final double mass;
-    private final double linearDamping;
-    private final double angularDamping;
-    private final double stability;
+    private final float mass;
+    private final float linearDamping;
+    private final float angularDamping;
+    private final float stability;
 
-    private BallKind(final double mass, final double linearDamping,
-            final double angularDamping, final double stability) {
+    private BallKind(final float mass, final float linearDamping,
+            final float angularDamping, final float stability) {
         this.mass = mass * MASS_SCALE;
         this.linearDamping = linearDamping;
         this.angularDamping = angularDamping;
         this.stability = stability * STABILITY_SCALE;
     }
 
-    public double getAngularDamping() {
+    public float getAngularDamping() {
         return angularDamping;
     }
 
-    public double getLinearDamping() {
+    public float getLinearDamping() {
         return linearDamping;
     }
 
-    public double getMass() {
+    public float getMass() {
         return mass;
     }
 
-    public double getStability() {
+    public float getStability() {
         return stability;
     }
 }
